base functions test done fix bug with pressBack
