+++ app/src/androidTest/java/com/strider/taskmanager/base/T11CreateTaskWithStatusAndPriorityTest.kt
@@ -1,4 +1,4 @@
-package com.strider.taskmanager
+package com.strider.taskmanager.base
 
 import androidx.test.espresso.Espresso
 import androidx.test.espresso.action.ViewActions
@@ -6,6 +6,7 @@ import androidx.test.espresso.assertion.ViewAssertions
 import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.strider.taskmanager.R
 import com.strider.taskmanager.enums.Priority
 import com.strider.taskmanager.enums.Status
 import com.strider.taskmanager.ui.MainActivity
