+++ app/src/androidTest/java/com/strider/taskmanager/base/T12CheckSaveTaskAfterBackPressedTest.kt
@@ -1,16 +1,15 @@
-package com.strider.taskmanager
+package com.strider.taskmanager.base
 
 import androidx.test.espresso.Espresso
 import androidx.test.espresso.Espresso.closeSoftKeyboard
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions
 import androidx.test.espresso.action.ViewActions.click
-import androidx.test.espresso.assertion.ViewAssertions
 import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.strider.taskmanager.R
 import com.strider.taskmanager.enums.Priority
 import com.strider.taskmanager.enums.Status
 import com.strider.taskmanager.ui.MainActivity
