+++ app/src/androidTest/java/com/strider/taskmanager/base/T13SwipeRightToDeleteTaskTest.kt
@@ -1,4 +1,4 @@
-package com.strider.taskmanager
+package com.strider.taskmanager.base
 
 import androidx.test.espresso.Espresso.*
 import androidx.test.espresso.action.ViewActions.*
@@ -7,8 +7,7 @@ import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.strider.taskmanager.enums.Priority
-import com.strider.taskmanager.enums.Status
+import com.strider.taskmanager.R
 import com.strider.taskmanager.ui.MainActivity
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import org.junit.Before
@@ -21,7 +20,7 @@ import kotlin.random.Random
 
 @ExperimentalCoroutinesApi
 @RunWith(AndroidJUnit4::class)
-class T10CreateTaskTest {
+class T13SwipeRightToDeleteTaskTest {
 
     @get:Rule
     var activityRule: ActivityScenarioRule<MainActivity> =
@@ -36,11 +35,18 @@ class T10CreateTaskTest {
         taskDescription = Random.nextLong().toString() + Random.nextLong().toString()
     }
 
-    // создаем задачу и проверяем её отображение
+    // сохранение задачи после нажатия системной кнопки назад
     @Test
-    fun t10CreateTaskTest() {
+    fun t13SwipeRightToDeleteTaskTest() {
         createTask()
         saveAndCheckCreatedTaskTest()
+
+        onView(withText(taskName)).perform(swipeRight())
+        Thread.sleep(2000)
+        onView(withText(taskName)).check(doesNotExist())
+        onView(withText(R.string.undo_upper_text)).check(matches(isDisplayed())).perform(click())
+        Thread.sleep(500)
+        onView(withText(taskName)).check(matches(isDisplayed()))
     }
 
     // создание задачи
