+++ app/src/main/java/com/strider/taskmanager/ui/MainActivity.kt
@@ -6,11 +6,9 @@ import androidx.activity.viewModels
 import androidx.databinding.DataBindingUtil
 import androidx.navigation.NavController
 import androidx.navigation.findNavController
-import androidx.navigation.get
 import androidx.navigation.ui.setupActionBarWithNavController
 import com.strider.taskmanager.R
 import com.strider.taskmanager.databinding.ActivityMainBinding
-import com.strider.taskmanager.ui.home.HomeFragment
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 
@@ -21,6 +19,7 @@ class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
     private lateinit var navController: NavController
     private var listener: OnBackPressedListener? = null
+    private val viewModel by viewModels<MainViewModel>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         setTheme(R.style.Theme_Custom)
@@ -40,8 +39,12 @@ class MainActivity : AppCompatActivity() {
         this.listener = listener
     }
 
+    fun setActionSave() {
+        viewModel.isTaskSaved = true
+    }
+
     override fun onBackPressed() {
-        if (navController.currentDestination?.id == R.id.taskDetailsFragment) {
+        if (navController.currentBackStackEntry?.destination?.id == R.id.taskDetailsFragment) {
             listener?.onBackPressed()
         } else {
             super.onBackPressed()
